Added afterUpdateQuery / afterInsertQuery

+++ src/Model.php
@@ -727,7 +727,9 @@ class Model implements \ArrayAccess, \IteratorAggregate
                 }
             }
 
-            $this->hook('beforeInsert', [&$data]);
+            if ($this->hook('beforeInsert', [&$data]) === false) {
+                return $this;
+            }
 
             // Collect all data of a new record
             $this->id = $this->persistence->insert($this, $data);
@@ -839,7 +841,9 @@ class Model implements \ArrayAccess, \IteratorAggregate
 
             return $this;
         } elseif ($this->loaded()) {
-            $this->hook('beforeDelete', [$this->id]);
+            if ($this->hook('beforeDelete', [$this->id])===false) {
+                return $this;
+            }
             $this->persistence->delete($this, $this->id);
             $this->hook('afterDelete', [$this->id]);
             $this->unload();
